+++ product-sdk/packages/host/src/index.ts
@@ -39,7 +39,6 @@ export {
     createHostPreimageManager,
     requestResourceAllocation,
     createProofAuthorized,
-    formatHostError,
     // Hex helpers
     toHex,
     fromHex,
@@ -47,7 +46,6 @@ export {
 export type {
     TruApi,
     HexString,
-    HostErrorPayload,
     PreimageManager,
     ResultAsync,
     AllocatableResource,
@@ -58,7 +56,14 @@ export type {
 // Result type + typed host errors (the throw→Result boundary)
 export { ok, err } from "./result.js";
 export type { Result } from "./result.js";
-export { HostError, HostUnavailableError, HostCallFailedError, isHostError } from "./errors.js";
+export {
+    HostError,
+    HostUnavailableError,
+    HostCallFailedError,
+    isHostError,
+    formatHostError,
+} from "./errors.js";
+export type { HostErrorPayload } from "./errors.js";
 
 // Accounts — host wallet accounts, product accounts, Ring VRF, and signers.
 export { getAccountsProvider } from "./accounts.js";
